Move the github package into the util/testhelpers package.
Even though the functionality that is provided thorugh this packaage
could be used in produciton code. It currently isn't. Since it's only
currently being used from test code, we can nest it under the
util/testhelpers package.
